Add files via upload
this tableau book includes the last two exercises 10-12, where we need to create a bucket to categorise the housing depending on the year of construction. We have later created the graphs to get a visual of the housing price per category and also created another graph with additional information, including variables such as bedrooms, conditions, grate, sqft (above, living, basement) among the ones that we have considered as relevant. 

irrespective of later discussion with colleagues on how to organise the information.